requirements.txt
PyYAML>=6.0
GitPython>=3.1.0
PyGithub>=2.1.1
dpath>=2.1.0
pytest>=7.0.0
pytest-cov>=4.0.0
